pypy-fastapi-container-instance/api/requirements.txt
fastapi
uvicorn
tortoise-orm
mimesis
aiosqlite